# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/items:
    get:
      tags:
        - Items
      summary: List Items
      description: Get a paginated list of Items with optional filtering and sorting
      operationId: listItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Items
      summary: Create Item
      description: Create a new Item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/item_groups:
    get:
      tags:
        - ItemGroups
      summary: List ItemGroups
      description: Get a paginated list of ItemGroups with optional filtering and sorting
      operationId: listItemGroups
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemGroupList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ItemGroups
      summary: Create ItemGroup
      description: Create a new ItemGroup
      operationId: createItemGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemGroupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Item:
      type: object
      required:
        - id
        - company_id
        - item_code
        - name
        - item_group_id
        - default_uom_id
        - item_type
        - is_sales_item
        - is_purchase_item
        - is_stock_item
        - has_variants
        - is_taxable
        - tags
        - data
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_code:
          type: string
          maxLength: 60
        name:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
          maxLength: 2000
        barcode:
          nullable: true
          type: string
          maxLength: 60
        brand_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          type: string
          format: uuid
        default_uom_id:
          type: string
          format: uuid
        item_type:
          type: string
        is_sales_item:
          type: boolean
        is_purchase_item:
          type: boolean
        is_stock_item:
          type: boolean
        has_variants:
          type: boolean
        hsn_code:
          nullable: true
          type: string
          maxLength: 20
        sni:
          nullable: true
          type: string
          maxLength: 30
        is_taxable:
          type: boolean
        weight_per_unit:
          nullable: true
          type: string
        shelf_life_days:
          nullable: true
          type: integer
          format: int32
        tags:
          type: object
        data:
          type: object
        status:
          type: string
        metadata:
          type: object
...
    CreateItemRequest:
      type: object
      required:
        - company_id
        - item_code
        - name
        - item_group_id
        - default_uom_id
        - item_type
        - is_sales_item
        - is_purchase_item
        - is_stock_item
        - has_variants
        - is_taxable
        - tags
        - data
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        item_code:
          type: string
          maxLength: 60
        name:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
          maxLength: 2000
        barcode:
          nullable: true
          type: string
          maxLength: 60
        brand_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          type: string
          format: uuid
        default_uom_id:
          type: string
          format: uuid
        item_type:
          type: string
        is_sales_item:
          type: boolean
        is_purchase_item:
          type: boolean
        is_stock_item:
          type: boolean
        has_variants:
          type: boolean
        hsn_code:
          nullable: true
          type: string
          maxLength: 20
        sni:
          nullable: true
          type: string
          maxLength: 30
        is_taxable:
          type: boolean
        weight_per_unit:
          nullable: true
          type: string
        shelf_life_days:
          nullable: true
          type: integer
          format: int32
        tags:
          type: object
        data:
          type: object
        status:
          type: string
        metadata:
          type: object
...
    ItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ItemGroup:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - is_group
        - level
        - sort_order
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        parent_id:
          nullable: true
          type: string
          format: uuid
        is_group:
          type: boolean
        level:
          type: integer
          format: int32
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateItemGroupRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - is_group
        - level
        - sort_order
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        parent_id:
          nullable: true
          type: string
          format: uuid
        is_group:
          type: boolean
        level:
          type: integer
          format: int32
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    ItemGroupList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemGroup'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []